+++ CHANGELOG.md
@@ -14,6 +14,8 @@
   - If underwater, marines will equip a pressure suit or MK50 to breathe infinitely.
     - Marines will only equip these armors if they have a full set of either armor type.
 - Marines can now be given potion items.
+  - If on fire, marines will now consider drinking a fire resistance potion.
+  - If underwater, marines will now consider drinking a water breathing potion if they are close to drowning.
 
 ## ♻️ Changes
 - N/A